+++ src/grandorgue/combinations/control/GOCombinationControllerProxy.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2026 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -13,11 +13,10 @@ void GOCombinationControllerProxy::PushGeneral(
     p_controller->PushGeneral(cmb, pButtonToLight);
 }
 
-void GOCombinationControllerProxy::PushDivisional(
-  GODivisionalCombination &cmb,
-  unsigned startManual,
-  unsigned cmbManual,
-  GOButtonControl *pButtonToLight) {
+void GOCombinationControllerProxy::ProcessPushDivisional(
+  unsigned startManualIndex,
+  std::function<std::pair<GOButtonControl *, GODivisionalCombination *>(
+    unsigned)> findManualDivisional) {
   if (p_controller)
-    p_controller->PushDivisional(cmb, startManual, cmbManual, pButtonToLight);
+    p_controller->ProcessPushDivisional(startManualIndex, findManualDivisional);
 }