+++ src/tests/GOTestExe.cpp
@@ -11,6 +11,7 @@
 
 #include "common/GOTestCollection.h"
 #include "testing/GOTestNameMap.h"
+#include "testing/combinations/GOTestDivisionalSetter.h"
 #include "testing/loader/GOTestOrganReader.h"
 #include "testing/model/GOTestDrawStop.h"
 #include "testing/model/GOTestOrganModel.h"
@@ -47,6 +48,7 @@ int main(int argc, char *argv[]) {
   }
 
   /* Instantiate all the test classes here */
+  GOTestDivisionalSetter testDivisionalSetter;
   GOTestOrganReader testOrganReader;
   GOTestDrawStop testDrawStop;
   GOTestOrganModel testOrganModel;